task: Add ParseTowArgs and SetupStorage
