+++ task/copy_file_generated.go
@@ -18,7 +18,7 @@ type CopyFileTask struct {
 	types.Pool
 	types.Storage
 	types.Path
-	types.ObjectKey
+	types.Key
 
 	// Runtime value
 	types.Todo
@@ -41,7 +41,7 @@ func initCopyFileTask(task types.Todoist) (t *CopyFileTask, o *CopyTask) {
 	t.SetPool(o.GetPool())
 	t.SetStorage(o.GetStorage())
 	t.SetPath(o.GetPath())
-	t.SetObjectKey(o.GetObjectKey())
+	t.SetKey(o.GetKey())
 	return
 }
 
@@ -51,7 +51,7 @@ type CopyLargeFileTask struct {
 	types.Pool
 	types.Size
 	types.Path
-	types.ObjectKey
+	types.Key
 	types.Storage
 
 	// Runtime value
@@ -80,7 +80,7 @@ func initCopyLargeFileTask(task types.Todoist) (t *CopyLargeFileTask, o *CopyFil
 	t.SetPool(o.GetPool())
 	t.SetSize(o.GetSize())
 	t.SetPath(o.GetPath())
-	t.SetObjectKey(o.GetObjectKey())
+	t.SetKey(o.GetKey())
 	t.SetStorage(o.GetStorage())
 	return
 }
@@ -90,7 +90,7 @@ type CopyPartialFileTask struct {
 	// Inherited value
 	types.Pool
 	types.Path
-	types.ObjectKey
+	types.Key
 	types.Storage
 	types.UploadID
 	types.WaitGroup
@@ -118,7 +118,7 @@ func initCopyPartialFileTask(task types.Todoist) (t *CopyPartialFileTask, o *Cop
 	t = &CopyPartialFileTask{}
 	t.SetPool(o.GetPool())
 	t.SetPath(o.GetPath())
-	t.SetObjectKey(o.GetObjectKey())
+	t.SetKey(o.GetKey())
 	t.SetStorage(o.GetStorage())
 	t.SetUploadID(o.GetUploadID())
 	t.SetWaitGroup(o.GetWaitGroup())
@@ -130,7 +130,7 @@ type CopySmallFileTask struct {
 	// Inherited value
 	types.Pool
 	types.Path
-	types.ObjectKey
+	types.Key
 	types.Storage
 	types.Size
 
@@ -155,7 +155,7 @@ func initCopySmallFileTask(task types.Todoist) (t *CopySmallFileTask, o *CopyFil
 	t = &CopySmallFileTask{}
 	t.SetPool(o.GetPool())
 	t.SetPath(o.GetPath())
-	t.SetObjectKey(o.GetObjectKey())
+	t.SetKey(o.GetKey())
 	t.SetStorage(o.GetStorage())
 	t.SetSize(o.GetSize())
 	return
